# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Agent AI Platform API
  version: "1.0.0"
  description: >
    Modular Monolith · Hexagonal · Plugin · Event-Driven.
    Auth: Firebase ID Token (local JWT verify). Tenant (workspace_id) derived from identity, never client-supplied.
    Errors: RFC 9457 Problem Details. Pagination: cursor-based. JSON: snake_case.
    All collection endpoints are wrapped as { data: [...], meta: { next_cursor, limit } } (API-04);
    bounded lists set next_cursor: null. Every response returns X-Correlation-Id.
    The interactive WebSocket endpoint /api/v1/ws is out of scope for OpenAPI and is specified in 03-api-spec.md §3.2.
servers:
  - url: /api/v1
security:
  - firebaseBearer: []
...
paths:
...
  /conversations/{id}/files:
    get:
      tags: [conversations]
      operationId: listConversationFiles
      description: "The files this thread's retrieval is PINNED to. Files themselves stay workspace-level (uploaded once, indexed once, reusable) — a pin is a retrieval SCOPE, not ownership: with pins, `rag_agent` searches only the documents built from these files; with none, it searches the whole workspace corpus, which is every existing thread's behaviour and stays the default. Returns references (`file_id`) and not file metadata, because the pin belongs to `conversations` and the name/size/status belong to `GET /files` — the client joins them by id, exactly as it already joins knowledge status. The set is bounded, so the page carries every pin and `meta.next_cursor` is always null. Permission `conversations:read`."
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ConversationFilePage" } } } }
        default: { $ref: "#/components/responses/Problem" }
    post:
      tags: [conversations]
      operationId: pinConversationFile
      description: "Pin one file into this thread's retrieval scope. The file must exist and be readable (`status=ready`, INV-F2) — pinning an unknown, quarantined, half-uploaded or deleted file is 422, never a stored reference to something retrieval can never reach. Idempotent: re-pinning returns 201 with the ORIGINAL `pinned_at`, so a double-click cannot manufacture a second row. Soft-deleted thread ⇒ 409, not 404. A pinned file that is later deleted is not an error — retrieval simply finds nothing for it, the same as a file that was never indexed. Permission `conversations:write`, the same as pinning the model: both configure how the thread answers."
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ConversationFileIn" } } } }
      responses:
        "201": { description: Pinned, content: { application/json: { schema: { $ref: "#/components/schemas/ConversationFileOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /conversations/{id}/files/{file_id}:
    delete:
      tags: [conversations]
      operationId: unpinConversationFile
      description: "Remove one pin. Deletes NO file and no index — it only widens this thread's retrieval scope back out, and unpinning the last one returns it to the whole workspace corpus. Idempotent (204 whether or not the pin was there); soft-deleted thread ⇒ 409, not 404. Permission `conversations:write`, not `conversations:delete`: nothing is destroyed."
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { name: file_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Unpinned }
        default: { $ref: "#/components/responses/Problem" }
...
  /files:
    get:
      tags: [files]
      operationId: listFiles
      parameters: [ { $ref: "#/components/parameters/SpaceId" }, { $ref: "#/components/parameters/Limit" }, { $ref: "#/components/parameters/Cursor" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/FilePage" } } } }
        default: { $ref: "#/components/responses/Problem" }
    post:
      tags: [files]
      operationId: registerFile
      description: Register an upload; returns a presigned PUT URL (MinIO).
      parameters: [ { $ref: "#/components/parameters/IdempotencyKey" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/FileRegisterIn" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/FileRegisterOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /files/{id}/complete:
    post:
      tags: [files]
      operationId: completeFile
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      requestBody: { required: false, content: { application/json: { schema: { $ref: "#/components/schemas/FileCompleteIn" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/FileOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /files/{id}:
    get:
      tags: [files]
      operationId: getFile
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/FileOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
    patch:
      tags: [files]
      operationId: renameFile
      description: "Rename a file — the ONLY mutable field it has. The bytes, the `content_type`, the `storage_key` and the `size_bytes` are all fixed at registration, and nothing here touches storage or the index: `knowledge.documents` keys on `file_id` and the Qdrant payload carries no name, so a rename costs no re-index and changes no answer. **The extension is immutable**: a name whose extension matches the current one (case-insensitively) is stored as sent; a name with NO extension inherits the current one; a name with a DIFFERENT extension is 422. The extension is the user-visible claim about bytes that cannot change, so letting it change would make the displayed name lie — and hand a download the wrong OS handler. Directory components are stripped and control characters refused (the same `FileName` sanitation as registration). Renaming to the current name is a no-op that still answers 200, WITHOUT bumping `updated_at` or the version. Any status may be renamed — fixing the name of a half-uploaded or quarantined file is exactly when it is wanted — but a soft-deleted file is 409, like every other write. Permission `files:write`, the same as registering."
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/FileRenameIn" } } } }
      responses:
        "200": { description: Renamed, content: { application/json: { schema: { $ref: "#/components/schemas/FileOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
    delete:
      tags: [files]
      operationId: deleteFile
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses:
        "204": { description: Soft-deleted }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    firebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    Limit: { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Cursor: { name: cursor, in: query, required: false, schema: { type: string } }
...
    SpaceId: { name: space_id, in: query, required: true, schema: { type: string, format: uuid } }
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: false, schema: { type: string } }
  headers:
    CorrelationId:
      description: "Request correlation id, echoed on every response (generated if absent). See 03 §0."
      schema: { type: string }
  responses:
    Problem:
      description: RFC 9457 Problem Details
      headers:
        X-Correlation-Id: { $ref: "#/components/headers/CorrelationId" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, correlation_id]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: "stable machine code, e.g. files.too_large" }
        correlation_id: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    PageMeta:
      type: object
      properties: { next_cursor: { type: string, nullable: true, description: "opaque cursor; null on the last (or unbounded) page" }, limit: { type: integer } }
...
    ConversationFileIn:
      type: object
      required: [file_id]
      properties: { file_id: {type: string, format: uuid} }
    ConversationFileOut:
      type: object
      properties:
        file_id: {type: string, format: uuid}
        pinned_at: {type: string, format: date-time, description: "UTC ISO-8601, trailing Z (DAT-06)", pattern: "Z$"}
    ConversationFilePage:
      type: object
      properties: { data: { type: array, items: { $ref: "#/components/schemas/ConversationFileOut" } }, meta: { $ref: "#/components/schemas/PageMeta" } }
...
    FileRegisterIn:
      type: object
      required: [space_id, name, content_type, size_bytes]
      properties: { space_id: {type: string, format: uuid, description: "The space this file belongs to. A file never moves between spaces, so this is the only moment its owner can be stated; it is also what the 1 GiB quota is charged against."}, name: {type: string, maxLength: 255}, content_type: {type: string}, size_bytes: {type: integer, minimum: 0} }
    FileRegisterOut:
      type: object
      properties: { file_id: {type: string, format: uuid}, upload_url: {type: string}, expires_in: {type: integer} }
    FileCompleteIn:
      type: object
      properties: { checksum: {type: string, nullable: true} }
    FileRenameIn:
      type: object
      required: [name]
      properties: { name: {type: string, maxLength: 255} }
    FileOut:
      type: object
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        content_type: {type: string}
        size_bytes: {type: integer}
        status: {type: string, enum: [uploaded, scanning, ready, quarantined]}
        download_url: {type: string, nullable: true}
        created_at: {type: string, format: date-time, description: "UTC ISO-8601, trailing Z (DAT-06)", pattern: "Z$"}
    FilePage:
      type: object
      properties: { data: { type: array, items: { $ref: "#/components/schemas/FileOut" } }, meta: { $ref: "#/components/schemas/PageMeta" } }
...